# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CookieNet API",
        "version": "1.0"
    },
    "paths": {
        "/": {},
        "/auth/sign-in": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "operationId": "88696b680d5e73bbbbfbc88b2ec80491",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "banana@paradise.net",
                                    "password": "123qweasd"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Create a new user",
                "operationId": "8060ba7cb59324131da8151d9518ca60",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Kiwi",
                                    "email": "banana@paradise.net",
                                    "password": "123qweasd"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/users/self": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get self user",
                "operationId": "134c15e26169315386829e3d42d18044",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}